requirements.txt
openai-whisper
transformers==4.36.2
torch==2.2.2
spacy
sentence-transformers
elasticsearch
opensearch-py
mlflow
apache-airflow
streamlit
pydub
python-dotenv
